backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "visquery-backend"
version = "0.1.0"
description = "Visquery backend — architectural precedent search API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic-settings>=2.2.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.29",
    "alembic>=1.13.1",
    "redis>=5.0.3",
    "rq>=1.16.2",
    "open-clip-torch>=2.24.0",
    # torch is declared without CUDA extras so pip pulls CPU-only wheels
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "faiss-cpu>=1.8.0",
    "Pillow>=10.3.0",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "anthropic>=0.25.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.6",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
